fix(collection-overview): fix undefined user in collection overview
