+++ src/collection/graphql.ts
@@ -194,6 +194,14 @@ export const GET_COLLECTIONS_BY_OWNER = gql`
 				stamboek
 				updated_at
 				user_id
+				user: usersByuserId {
+					first_name
+					last_name
+					role {
+						id
+						label
+					}
+				}
 			}
 			is_public
 			external_id
