+++ src/collection/views/CollectionDetail.tsx
@@ -32,6 +32,7 @@ import { RouteParts } from '../../constants';
 import ControlledDropdown from '../../shared/components/ControlledDropdown/ControlledDropdown';
 import { DataQueryComponent } from '../../shared/components/DataComponent/DataQueryComponent';
 import DeleteObjectModal from '../../shared/components/modals/DeleteObjectModal';
+import { renderAvatar } from '../../shared/helpers/formatters/avatar';
 import { formatDate } from '../../shared/helpers/formatters/date';
 import {
 	generateAssignmentCreateLink,
@@ -45,7 +46,6 @@ import { DELETE_COLLECTION, GET_COLLECTION_BY_ID } from '../graphql';
 import { ContentTypeString } from '../types';
 
 import './CollectionDetail.scss';
-import { renderAvatar } from '../helpers';
 
 interface CollectionDetailProps extends RouteComponentProps {}
